fix(file-upload): moved file-validators and updated examples
