+++ packages/components/core/forms/index.ts
@@ -1,4 +1,3 @@
-export * from './file-validators';
 export * from './forms-module';
 export * from './forms.directive';
 export * from './validators';
