+++ packages/docs-examples/components/file-upload/file-upload-single-validation-reactive-forms-overview/file-upload-single-validation-reactive-forms-overview-example.ts
@@ -3,7 +3,7 @@ import { FormControl, FormGroup, ReactiveFormsModule } from '@angular/forms';
 import { FileValidators } from '@koobiq/components/core';
 import { KbqFileUploadModule } from '@koobiq/components/file-upload';
 import { KbqFormFieldModule } from '@koobiq/components/form-field';
-import { KbqIcon } from '@koobiq/components/icon';
+import { KbqIconModule } from '@koobiq/components/icon';
 
 const MAX_FILE_SIZE = 5 * 2 ** 20;
 
@@ -19,17 +19,17 @@ const MAX_FILE_SIZE = 5 * 2 ** 20;
                 class="layout-margin-bottom-s"
                 formControlName="fileControl"
             >
-                @if (!this.fileControl.errors) {
+                @if (!formGroup.get('fileControl')?.errors) {
                     <i kbq-icon="kbq-file-o_16"></i>
                 }
-                @if (this.fileControl.errors) {
+                @if (formGroup.get('fileControl')?.errors) {
                     <i kbq-icon="kbq-exclamation-triangle_16"></i>
                 }
 
-                <kbq-hint>Размер файла не должен быть более 5 МБ</kbq-hint>
+                <kbq-hint>The file size should not exceed 5 MB</kbq-hint>
 
-                @if (this.fileControl.hasError('maxFileSize')) {
-                    <kbq-hint color="error">Размер файла не должен превышать лимит.</kbq-hint>
+                @if (formGroup.get('fileControl')?.hasError('maxFileSize')) {
+                    <kbq-hint color="error">The file size must not exceed the limit</kbq-hint>
                 }
             </kbq-file-upload>
         </form>
@@ -37,7 +37,7 @@ const MAX_FILE_SIZE = 5 * 2 ** 20;
     imports: [
         KbqFileUploadModule,
         KbqFormFieldModule,
-        KbqIcon,
+        KbqIconModule,
         ReactiveFormsModule
     ]
 })
@@ -45,8 +45,4 @@ export class FileUploadSingleValidationReactiveFormsOverviewExample {
     formGroup = new FormGroup({
         fileControl: new FormControl(null, FileValidators.maxFileSize(MAX_FILE_SIZE))
     });
-
-    get fileControl() {
-        return this.formGroup.get('fileControl') as FormControl;
-    }
 }
